fixed isHomePage check
Before, the homepage was checked via:

    $this->getUrl('') == $this->getUrl('*/*/*', array('_current'=>true, '_use_rewrite'=>true));

Although this is the "official" code as in `Mage_Page_Block_Html_Header::getIsHomePage`, this is buggy. It returns false if you hit the homepage with a parameter like `https://www.site.com/?__from_store=de`. The full action name should be a better way of checking I think.