getDiscontinuedCategoryUrl now checks to see if category parent exists
I had an issue on a site where the category no longer existed, this caused a 502 error when google crawled the old category, as it caused an infinite loop since the else statement calls it's own method. 

It should be returning a 404. 

I've changed line 60 to an if else statement so if a category does not have a parent Id, it doesn't call it's own method again.